fixed issue in controller

+++ src/main/java/com/example/idcard/service/PersonServiceImpl.java
@@ -4,6 +4,8 @@ import com.example.idcard.dto.PersonDto;
 import com.example.idcard.model.Person;
 import com.example.idcard.repository.PersonRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,22 +17,36 @@ public class PersonServiceImpl implements PersonService {
     PersonRepo personRepo;
 
     @Override
-    public void createPerson(Person person) {
+    public ResponseEntity createPerson(Person person) {
         personRepo.save(person);
+        return ResponseEntity.status(HttpStatus.CREATED).body("New Person is created!");
+
     }
 
     @Override
-    public void updatePerson(Person person,int id) {
-        Person person1 = personRepo.getById(id);
-        person1.setFatherName(person.getFatherName());
-        person1.setFirstName(person.getFirstName());
-        person1.setLastName(person.getLastName());
-        personRepo.save(person1);
+    public ResponseEntity updatePerson(Person person,int id) {
+        if(exists(id)){
+            Person person1 = personRepo.getById(id);
+            person1.setFatherName(person.getFatherName());
+            person1.setFirstName(person.getFirstName());
+            person1.setLastName(person.getLastName());
+            personRepo.save(person);
+
+            return ResponseEntity.status(HttpStatus.OK).body("Person with id:"+id+" is updated!");
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("There is no person with id: "+id);
+        }
+
     }
 
     @Override
-    public void deletePerson(int id) {
-        personRepo.deleteById(id);
+    public ResponseEntity deletePerson(int id) {
+        if(exists(id)){
+            personRepo.deleteById(id);
+            return ResponseEntity.status(HttpStatus.OK).body("person with id: "+id+" is deleted");
+        }else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("There is no person with this id");
+        }
     }
 
 
@@ -40,8 +56,12 @@ public class PersonServiceImpl implements PersonService {
     }
 
     @Override
-    public PersonDto getPerson(int id) {
-        return personRepo.findById(id);
+    public ResponseEntity<Object> getPerson(int id) {
+        if(exists(id)){
+            return ResponseEntity.status(HttpStatus.OK).body(personRepo.findById(id));
+        }else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("There is no person with id: "+id);
+        }
     }
 
     @Override
